requirements.txt
uvicorn==0.11.3
pydantic==1.4
fastapi==0.54.1
PyJWT==1.7.1
PyYAML==5.3.1
